requirements.txt
openai
PyMuPDF
python-docx
